refactor(Notice): AiSummaryResult를 application 계층으로 이동
- AiPort(application)가 infra.summarization의 타입을 직접 참조하던
  의존 방향 위반을 해소
- Port가 이제 application 계층 내부 타입만 참조하도록 정리
