+++ backend/src/main/java/org/cathori/backend/notice/application/AiPort.java
@@ -1,8 +1,6 @@
 package org.cathori.backend.notice.application;
 
 
-import org.cathori.backend.notice.infra.summarization.AiSummaryResult;
-
 import java.util.List;
 
 public interface AiPort {
