+++ backend/src/main/java/org/cathori/backend/notice/infra/summarization/GeminiAiAdapter.java
@@ -1,6 +1,7 @@
 package org.cathori.backend.notice.infra.summarization;
 
 import org.cathori.backend.notice.application.AiPort;
+import org.cathori.backend.notice.application.AiSummaryResult;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 import org.springframework.web.client.RestClient;
