+++ backend/src/main/java/org/cathori/backend/notice/infra/summarization/NoticeSummaryUpdater.java
@@ -1,6 +1,7 @@
 package org.cathori.backend.notice.infra.summarization;
 
 import lombok.RequiredArgsConstructor;
+import org.cathori.backend.notice.application.AiSummaryResult;
 import org.cathori.backend.notice.model.Notice;
 import org.cathori.backend.notice.model.NoticeRepository;
 import org.springframework.stereotype.Component;
