+++ backend/src/main/java/org/cathori/backend/notice/model/Notice.java
@@ -6,7 +6,7 @@ import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.cathori.backend.notice.application.crawling.CrawledNotice;
-import org.cathori.backend.notice.infra.summarization.AiSummaryResult;
+import org.cathori.backend.notice.application.AiSummaryResult;
 import tools.jackson.core.JacksonException;
 import tools.jackson.databind.ObjectMapper;
